Add AT&T-style instruction mnemonic display
