+++ op-common.c
@@ -41,21 +41,3 @@ void print_column (char * s, byte_t w)
 		d++;
 		}
 	}
-
-
-// Print a relative number with optional sign prefix
-
-void print_rel (byte_t prefix, short rel)
-	{
-	if (rel >= 0)
-		{
-		if (prefix) putchar ('+');
-		}
-	else
-		{
-		putchar ('-');
-		rel = -rel;
-		}
-
-	printf ("%hXh", (word_t) rel);
-	}
