+++ op-id-name.h
@@ -4,4 +4,4 @@
 #include "op-common.h"
 
 
-char *op_id_to_name (word_t op_id);
+char *op_id_to_name (word_t op_id, word_t lower);
